add: a configurable HTTP transport connect timeout
* A new env variable HONEYCOMB_CONNECT_TIMEOUT can be used to customize the HTTP
transport connection timeout for the underlying libhoney-go client so that when
there are connectivity issues, batch upload will fail faster and not hang for the
full duration of the batch upload timeout. The default connection timeout is set
to 3 seconds if not specified. This closes out a portion of issue #101, as it's
only handling connect timeout for now. Limiting network connection timeouts
is critical for a Lambda extension as it's possible for an extension to extend
the duration of a function invocation and result in function timeouts. This has
real world impacts beyond just costs. This scenario can easily result in
overall Lambda account concurrency limits being exceeded, which will prevent further
function executions from running (essentially a denial of service). Furthermore,
when function timeouts occur, this causes the underyling Lambda sandbox to terminate
resulting in performance impact in the form of cold starts for all invocations.